requirements.txt
streamlit
pandas
openpyxl
langchain
langchain-openai
langchain-community
matplotlib
python-dotenv
